+++ bruges/petrophysics/__init__.py
@@ -1,3 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .petrophysics import gardner
+from .petrophysics import gardner, inverse_gardner
+from .petrophysics import velocity_to_density, density_to_velocity
+from .petrophysics import density_to_porosity, porosity_to_density
+from .petrophysics import velocity_to_slowness, slowness_to_velocity