+++ bruges/petrophysics/petrophysics.py
@@ -9,26 +9,53 @@ petrophysics.py
 """
 
 
-def gardner(vp, alpha=0.31, beta=0.25, fps=False):
+def gardner(vp, alpha=310, beta=0.25, fps=False):
     """
     Computes Gardner's density prediction from P-wave velocity.
 
     Args:
         vp (ndarray): P-wave velocity in m/s.
-        alpha (float): The factor, 0.31 for m/s and 0.23 for fps.
+        alpha (float): The factor, 310 for m/s and 230 for ft/s.
         beta (float): The exponent, usually 0.25.
         fps (bool): Set to true for FPS and the equation will use the typical
-            value for alpha.
+            value for alpha. Overrides value for alpha, so if you want to use
+            you own alpha, regardless of units, set this to False.
 
     Returns:
-        ndarray: RHOB estimate.
+        ndarray: RHOB estimate in kg/m^3.
     """
-    if fps:
-        alpha = 0.23
+    alpha = 230 if fps or alpha
     return alpha * vp ** beta
 
 
-def porosity_to_density(phi_rhob, rho_matrix, rho_fluid):
+def inverse_gardner(rho, alpha=310, beta=0.25, fps=False):
+    """
+    Computes Gardner's density prediction from P-wave velocity.
+
+    Args:
+        rho (ndarray): Density in kg/m^3.
+        alpha (float): The factor, 310 for m/s and 230 for fps.
+        beta (float): The exponent, usually 0.25.
+        fps (bool): Set to true for FPS and the equation will use the typical
+            value for alpha. Overrides value for alpha, so if you want to use
+            you own alpha, regardless of units, set this to False.
+
+    Returns:
+        ndarray: Vp estimate in m/s.
+    """
+    alpha = 230 if fps or alpha
+    exponent = 1 / beta
+    factor = 1 / alpha**exponent
+    return factor * rho**exponent
+
+
+velocity_to_density = gardner
+
+
+density_to_velocity = inverse_gardner
+
+
+def porosity_to_density(phi, rho_matrix, rho_fluid):
     """
     Get density from a porosity log. Typical values:
 
@@ -44,17 +71,17 @@ def porosity_to_density(phi_rhob, rho_matrix, rho_fluid):
     See wiki.aapg.org/Density-neutron_log_porosity.
 
     Args:
-        phi_rhob (ndarray): The density porosity log.
+        phi (ndarray): The porosity log.
         rho_matrix (float)
         rho_fluid (float)
 
     Returns:
-        Estimate of RHOB.
+        Estimate of bulk density, rho.
     """
-    return rho_matrix * (1 - phi_rhob) + rho_fluid * phi_rhob
+    return rho_matrix * (1 - phi) + rho_fluid * phi_rhob
 
 
-def density_to_porosity(density, rho_matrix, rho_fluid):
+def density_to_porosity(rho, rho_matrix, rho_fluid):
     """
     Get density from a porosity log. Typical values:
 
@@ -70,25 +97,28 @@ def density_to_porosity(density, rho_matrix, rho_fluid):
     See wiki.aapg.org/Density-neutron_log_porosity.
 
     Args:
-        density (ndarray): The density log or RHOB.
+        rho (ndarray): The bulk density log or RHOB.
         rho_matrix (float)
         rho_fluid (float)
 
     Returns:
         Estimate of porosity as a volume fraction.
     """
-    return (rho_matrix - density) / (rho_matrix - rho_fluid)
+    return (rho_matrix - rho) / (rho_matrix - rho_fluid)
 
 
 def slowness_to_velocity(slowness):
     """
-    Convert a DT or DTS log in µs per unit depth, to velocity in unit depth
+    Convert a slowness log in µs per unit depth, to velocity in unit depth
     per second.
 
     Args:
-        slowness (ndarray): A DT value or sequence of values.
+        slowness (ndarray): A value or sequence of values.
 
     Returns:
         ndarray: The velocity.
     """
     return 1e6 / np.array(slowness)
+
+
+velocity_to_slowness = slowness_to_velocity